requirements.txt
django==4.0.2
matplotlib==3.5.1
Pillow==9.0.1
numpy==1.21.6
numba==0.55.1